pyproject.toml
[project]
name = "hato"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.3",
    "numpy>=2.2.6",
    "ollama>=0.5.1",
    "openai>=1.82.1",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
    "seaborn>=0.13.2",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.11.12",
    "ty>=0.0.1a7",
    "types-seaborn>=0.13.2.20250516",
]
